Partially revamp icon adding

+++ src/constants.hpp
@@ -2,5 +2,5 @@
 
 // Idk if there's a function to get the color count
 namespace changing_icons::constants {
-    inline size_t constexpr COLOR_COUNT = 106;
+    inline int constexpr COLOR_COUNT = 106;
 }
